% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate.tracks}
\alias{aggregate.tracks}
\title{Compute Summary Statistics of Subtracks}
\usage{
\method{aggregate}{tracks}(x, measure, by = "subtracks", FUN = mean,
  subtrack.length = seq(1, (maxTrackLength(x) - 1)),
  max.overlap = max(subtrack.length), na.rm = FALSE, ...)
}
\arguments{
\item{x}{the tracks object whose subtracks are to be considered.
If a single track is given, it will be coerced to a tracks object
using \code{\link{wrapTrack}} (but note that this requires an explicit call
\code{aggregate.tracks}).}

\item{measure}{the measure that is to be computed on the subtracks.}

\item{by}{a string that indicates how grouping is performed. Currently, two
kinds of grouping are supported:
\itemize{
 \item{\code{"subtracks"}}{ Apply \code{measure} to all subtracks according to
the parameters \code{subtrack.length} and \code{max.overlap}.}
 \item{\code{"prefixes"}}{ Apply \code{measure} to all prefixes (i.e., subtracks starting
 from a track's initial position) according to the parameter \code{subtrack.length}.}
}}

\item{FUN}{a function that is to be used to aggregate the measures
of corresponding subtracks.
If a string is given, it is first matched to the following builtin values:
\itemize{
 \item{"mean.sd"}{ Outputs the mean and \eqn{mean - sd} as lower and
 \eqn{mean + sd} as upper bound}
 \item{"mean.se"}{ Outputs the mean and \eqn{mean - se} as lower and
 \eqn{mean + se} as upper bound}
 \item{"mean.ci.95"}{ Outputs the mean and upper and lower bound of the
 symmetric 95 percent confidence intervall}.
 \item{"mean.ci.99"}{ Outputs the mean and upper and lower bound of the
 symmetric 95 percent confidence intervall}.
 \item{"iqr"}{ Outputs the interquartile range, that is, the median, and the
 25-percent-quartile as a lower and and the 75-percent-quartile as an
 upper bound}
}
If the string is not equal to any of these, it is passed on to
\code{\link{match.fun}}}

\item{subtrack.length}{either a numeric value \eqn{i}, indicating that only
subtracks of exactly \eqn{i} segments are to be regarded, or a vector of all
the values \eqn{i} that shall be used. In particular, \code{subtrack.length=2}
corresponds to a "step-based analysis" (Beltman et al, 2009).}

\item{max.overlap}{Determines the amount by which the subtracks that are taken into
account can overlap. A maximum overlap of \code{max(subtrack.length)} will imply
that all subtracks are considered. For a maximum overlap of 0, only non-overlapping
subtracks are considered. A negative overlap can be used to ensure that only subtracks
a certain distance apart are considered. In general, for non-Brownian motion there will
be correlations between subsequent steps, such that a negative overlap may be necessary
to get a proper error estimate.}

\item{na.rm}{This is passed on to the builtin statistics like "mean.se" or
"mean".}

\item{...}{further arguments passed to or used by methods.}
}
\value{
a data frame with one row for every \eqn{i} in the set
specified by \code{subtrack.length}. The first column contains the values
of \eqn{i} and the remaining columns contain the values of the summary statistic
of the measure values of tracks having exactly \eqn{i} segments.
}
\description{
This is the main workhorse function to compute many common motility measures
such as mean square displacement, turning angle, and autocorrelation functions.
}
\details{
For the given \emph{tracks} object, a track measure is applied to all the
tracks' subtracks of a certain number of segments \eqn{i}.
On all the values obtained for a certain value of \eqn{i}, the given
statistic is computed, and the resulting values are returned.

For every number of segments \eqn{i} in the set defined by
\code{subtrack.length}, all subtracks of any track in the input
\emph{\code{tracks}} object, that consist of exactly \eqn{i} segments are
considered. The input \code{measure} is applied to the subtracks individually,
and the \code{statistic} is applied to the resulting values.
The summary statistics are returned for each value of \eqn{i} in a data frame.
}
\examples{
## A mean square displacement plot with error bars.
dat <- aggregate(TCells, squareDisplacement, FUN="mean.se")
with( dat ,{
  plot( mean ~ i, xlab="time step",
  	ylab="mean square displacement", type="l" )
  segments( i, lower, y1=upper )
} )
}
\references{
Joost B. Beltman, Athanasius F.M. Maree and Rob. J. de Boer (2009).
Analysing immune cell migration. \emph{Nature Reviews Immunology} \bold{9},
789--798. doi:10.1038/nri2638
}

