% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregateSubtrackMeasures}
\alias{aggregateSubtrackMeasures}
\title{Compute a Statistic on Measure Values of Subtracks of a Certain Length}
\usage{
aggregateSubtrackMeasures(tracks, measure, statistic = mean,
  subtrack.length = seq(1, (maxTrackLength(tracks) - 1)),
  prefixes.only = FALSE, max.overlap = max(subtrack.length),
  na.rm = FALSE)
}
\arguments{
\item{tracks}{the tracks object whose subtracks are to be considered. If a single track
is given, it will be coerced to a tracks object using \code{\link{wrapTrack}}.}

\item{measure}{the measure that is to be computed on the subtracks.}

\item{statistic}{the function that is to be used to aggregate the measures
of subtracks of the same number of segments. The function may return an
arbitrary number of values.
For the most common functions, a string can be given:
\itemize{
 \item{"mean.sd"}{ Outputs the mean and \eqn{mean - sd} as lower and
 \eqn{mean + sd} as upper bound}
 \item{"mean.se"}{ Outputs the mean and \eqn{mean - se} as lower and
 \eqn{mean + se} as upper bound}
 \item{"mean.ci.95"}{ Outputs the mean and upper and lower bound of the
 symmetric 95 percent confidence intervall}.
 \item{"mean.ci.99"}{ Outputs the mean and upper and lower bound of the
 symmetric 95 percent confidence intervall}.
 \item{"iqr"}{ Outputs the interquartil range, that is, the median, and the
 25-percent-quartile as a lower and and the 75-percent-quartile as an
 upper bound}
}}

\item{subtrack.length}{either a numeric value \eqn{i}, indicating that only
subtracks of exactly \eqn{i} segments are to be regarded, or a vector of all
the values \eqn{i} that shall be used.}

\item{prefixes.only}{logical. If TRUE, then only subtracks starting from a track's
first position are considered. Could be relevant for tracks for which the starting
position is indeed meaningful (instead of being simply the position at which observation
of a longer trajectory started).}

\item{max.overlap}{Determines the amount by which the subtracks that are taken into
account can overlap. A maximum overlap of \code{max(subtrack.length)} will imply
that all subtracks are considered. For a maximum overlap of 0, only non-overlapping
subtracks are considered. A negative overlap can be used to ensure that only subtracks
a certain distance apart are considered. In general, for non-Brownian motion there will
be correlations between subsequent steps, such that a negative overlap may be necessary
to get a proper error estimate.}

\item{na.rm}{This is passed on to the builtin statistic functions like "mean.se" or
"mean".}
}
\value{
Returns a data frame with one line for every \eqn{i} in the set
specified by \code{subtrack.length}. The first column contains the values
of \eqn{i}, the following columns contain the values of the agglomeration
function applied on the measure values of tracks of exactly \eqn{i} segments.
}
\description{
For a given \emph{tracks} object, a given measure is applied to all the
tracks' subtracks of a certain number of segments \eqn{i}.
On all the values obtained for a certain value of \eqn{i}, the given
statistic is computed, and all the resulting values are returned.
}
\details{
For every number of segments \eqn{i} in the set defined by
\code{subtrack.length}, all subtracks of any track in the input
\emph{\code{tracks}} object, that consist of exactly \eqn{i} segments are
regarded. The input \code{measure} is applied to the subtracks individually,
and the \code{statistic} is used to agglomerate the obtained values.
The return values of the agglomeration function are returned for each value
of \eqn{i} in a data frame.
}
\examples{
require(ggplot2)
dat <- aggregateSubtrackMeasures(TCells, displacement, "mean.se")
ggplot(dat, aes(x=i, y=mean)) + geom_errorbar(aes(ymin=lower, ymax=upper), width=.1)
}

