% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.tracks.list}
\alias{as.tracks.list}
\title{Convert from List to Tracks}
\usage{
\method{as.tracks}{list}(x, ...)
}
\arguments{
\item{x}{the list that is to be converted into a \emph{tracks} object.}

\item{...}{further arguments to be passed from or to other methods.}
}
\description{
"Converts" a list into a \code{tracks} object by settings its 
S3 class to \code{tracks}.
}

