% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{as.tracks}
\alias{as.tracks}
\title{Convert to tracks}
\usage{
as.tracks(x, ...)
}
\arguments{
\item{x}{the object to be coerced.}

\item{...}{further arguments to be passed on to the respective method.}
}
\value{
the coerced object of S3 class \code{tracks}.
}
\description{
Coerces \code{x} to a \code{tracks} object.
}
\details{
the S3 Method corresponding to \code{x}'s type (usually a list) is called.
}

