% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{asphericity}
\alias{asphericity}
\title{Track Aspericity}
\usage{
asphericity(track)
}
\arguments{
\item{track}{the track whose asphericity is to be computed. Note that the
track must be in a two- or threedimensional space, thus the data frames
containing the track's points must contain an \eqn{x} and \eqn{y} value.}
}
\value{
The track's asphericity, a vaule between \eqn{0} and \eqn{1}, will
be returned for two- or threedimensional tracks, otherwise \code{NA}.
}
\description{
Computes the asphericity of the track's points.
}
\details{
Computes the asphericity of a track, i.e. the scatter plot of
its points, by considering the length of its principal components,
see References. The value will be from the interval \eqn{[0,1]},
with high values indicating track straightness. The asphericity is a
measure for the track's straightness. Note that the asphericity is only
defined for tracks in two or three dimensions.
}
\references{
Zeinab Mokhtari, Franziska Mech, Carolin Zitzmann, Mike Hasenberg, Matthias Gunzer
and Marc Thilo Figge (2013), Automated Characterization and
Parameter--Free Classification of Cell Tracks Based on Local Migration
Behavior. \emph{PLoS ONE} \bold{8}(12), e80808. doi:10.1371/journal.pone.0080808
}

