% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boundingBox}
\alias{boundingBox}
\title{Bounding Box of a Tracks Object}
\usage{
boundingBox(x)
}
\arguments{
\item{x}{the \code{tracks} object whose bounding box is to be computed.}
}
\value{
Returns a matrix with two rows and \eqn{d} columns, where \eqn{d} is 
the number of dimensions of the tracks. The first row contains the minimum 
and the second row the maximum value of any track in the dimension given by 
the column.
}
\description{
Computes the minimum and maximum coordinates per dimension for all positions in a
given list of tracks. Assumes that all tracks have the same number of dimensions.
}

