% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-tracks.R
\name{brownianTrack}
\alias{brownianTrack}
\title{Simulate \code{nsteps} Steps of a Random Walk in \code{dim} Dimensions}
\usage{
brownianTrack(nsteps = 100, dim = 3)
}
\arguments{
\item{nsteps}{number of steps the track shall consist of.}

\item{dim}{the number of dimensions the track shall have.}
}
\value{
A data frame  containing in cell track with \code{nsteps} steps in
\code{dim} dimensions is returned.
}
\description{
Generates a random track with \code{nsteps} steps in \code{dim} dimensions.
}
\details{
In in every step an for each dimension, a normally distributed
value is added to the previous cell position.
}

