% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{c.tracks}
\alias{c.tracks}
\title{Join Track Datasets}
\usage{
\method{c}{tracks}(...)
}
\arguments{
\item{...}{the \emph{tracks} objects that are to be concatenated.}
}
\value{
A \emph{tracks} object containing all the tracks given in any of 
the input \emph{tracks} objects.
}
\description{
Concatenates several \emph{tracks} objects.
}
\details{
Joins the tracks (given as data frames) in the given \emph{tracks} 
objects into one tracks object, preserving the names of list elements and 
the ordering of the input objects as well as of the individual tracks in 
them.
}

