% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{clusterTracks}
\alias{clusterTracks}
\title{Cluster Tracks}
\usage{
clusterTracks(tracks, measures, scale = TRUE, ...)
}
\arguments{
\item{tracks}{the tracks that are to be clustered.}

\item{measures}{the measures according to which the tracks are to be
clustered.}

\item{scale}{logical indicating whether the measures values shall be scaled
using the function \code{\link[base]{scale}} before the clustering. Default
is TRUE.}

\item{...}{additional parameters to be passed to \code{\link[stats]{hclust}}.}
}
\value{
An object of class *hclust*, see \code{\link[stats]{hclust}}.
}
\description{
Clusters the tracks in a given \emph{tracks} object according to given
measures.
}
\details{
The measures are applied to each of the tracks in the given
\emph{tracks} object and according to the resulting values the tracks are
clustered using a hierarchical clustering (see \code{\link[stats]{hclust}}).
If \code{scale} is set to true, the values that have been determined by
applying the input measures to the input tracks are scaled to mean value
\eqn{0} and standard deviation \eqn{1} (per measure) before the clustering.
}

