% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{computeForEveryPrefix}
\alias{computeForEveryPrefix}
\title{Apply a Measure to Every Prefix of a Given Track}
\usage{
computeForEveryPrefix(track, measure, min.length = 1)
}
\arguments{
\item{track}{the track whose prefixes are to be regarded.}

\item{measure}{the measure that is to be applied.}

\item{min.length}{minimal number of points in regarded prefixes. Default is
\eqn{1}.}
}
\value{
Returns a vector with the values of the measure applied to each of
the given track's prefixes.
}
\description{
Computes a vector in which each position represents the value of the given
measure for a prefix of certain length of the input track.
}
\details{
The measure is applied to every prefix of the input track, starting
with the prefix that consists of \code{min.length} points. The resulting
values are returned in a vector, sorted in ascending order of the
corresponding prefix's length.
}

