% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{computeSubtracksOfISegments}
\alias{computeSubtracksOfISegments}
\title{All Subtracks of a Given Length}
\usage{
computeSubtracksOfISegments(track, i, overlap = i - 1)
}
\arguments{
\item{track}{the track whose subtracks shall be computed.}

\item{i}{integer indicating the desired subtrack length}

\item{overlap}{the number of segments that shall overlap in each two
subsequent subtracks.}
}
\value{
A list of data frames (a \emph{tracks} object), containing a data
frame for each of the track's subtracks that constists of exactly \eqn{i} single
segments.
}
\description{
For a given integer \eqn{i} and a track, this function returns a tracks
object of all subtracks that consist of exactly \eqn{i} subsegments.
}
\details{
Beginning with the track's starting point, for each point on
the track, this point and the subsequent \eqn{i} points are saved
in a data frame. All the data frames are returned as a list, constituting
an object of class \emph{tracks}.
}

