% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacement}
\alias{displacement}
\title{Track Displacement}
\usage{
displacement(x, from = 1, to = nrow(x))
}
\arguments{
\item{x}{the track whose displacement is to be computed.}

\item{from}{index of first row of the track.}

\item{to}{index of last row of the track.
Both `from` and `to` can be vectors and are useful to avoid extracting subtracks,
which is exploited by \code{\link{aggregate.tracks}}.}
}
\value{
The track's displacement, a nonnegative number.
}
\description{
Computes the Euclidean distance between the track's start and end points.
}

