% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacementRatio}
\alias{displacementRatio}
\title{Track Displacement Ratio}
\usage{
displacementRatio(x)
}
\arguments{
\item{x}{the track whose displacement ratio is to be computed.}
}
\value{
The track's displacement ratio, a value between \eqn{0} and \eqn{1}, 
will be returned. If the track has a maximal displacement of \eqn{0}, then 
1 is returned.
}
\description{
Computes a track's displacement ratio, i.d. the track's (current) 
displacement divided by its maximal displacement.
}
\references{
Zeinab Mokhtari, Franziska Mech, Carolin Zitzmann, Mike Hasenberg, Matthias Gunzer
and Marc Thilo Figge (2013), Automated Characterization and 
Parameter--Free Classification of Cell Tracks Based on Local Migration 
Behavior. \emph{PLoS ONE} \bold{8}(12), e80808. doi:10.1371/journal.pone.0080808
}
\seealso{
\code{\link{displacement}}, \code{\link{maxDisplacement}}
}

