% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{duration}
\alias{duration}
\title{Track Duration}
\usage{
duration(x)
}
\arguments{
\item{x}{a track.}
}
\value{
The track's duration, a nonnegative number.
}
\description{
Returns the duration of a track, i.e., the time elapsed between its first and last
position.
}

