% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{fractalDimension}
\alias{fractalDimension}
\title{Track Fractal Dimension}
\usage{
fractalDimension(track)
}
\arguments{
\item{track}{the track for which the fractal dimension is to be calculated.}
}
\value{
the number indicating the track's fractal dimension.
}
\description{
Computes the fractal dimension of a track using all track dimensions by the
box-count method.
}
\details{
The fractal dimension is estimated using the function from
\code{\link[fractaldim]{fd.estim.boxcount}} from the
\code{fractaldim} package. For Brownian motion, fractal dimension and Hurst exponent
(see \link{hurstExponent})
are related by the formula H=2-D. For non-Brownian motion, however, this relationship
need not hold. While the hurst exponent takes a global approach to the
track's properties, fractal dimension is a local approach to the track's properties
(Gneiting and Schlather, 2004).
}
\references{
Tillmann Gneiting and Martin Schlather (2004), Stochastic Models That Separate Fractal
Dimension and the Hurst Effect. \emph{SIAM Review} \bold{46}(2), 269--282.
doi:10.1137/S0036144501394387
}
\seealso{
\code{\link{hurstExponent}}
}

