% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getSubsetOfTracks}
\alias{getSubsetOfTracks}
\title{Extract tracks based on user defined properties}
\usage{
getSubsetOfTracks(tracks, measure, ll, ul)
}
\arguments{
\item{tracks}{the tracks from which subsetting is to be done on}

\item{measure}{the track measure for which the subsetting is to be based on}

\item{ll}{specifies the lower-limit of the allowable measure}

\item{ul}{specifies the upper-limit of the allowable measure}
}
\description{
Given a tracks object, extract a subset based on upper and lower bounds of a certain
measure. For instance, extract all tracks with a certain minimum length.
}

