% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getTracks}
\alias{getTracks}
\title{Get a Subset of Tracks}
\usage{
getTracks(tracks, ids)
}
\arguments{
\item{tracks}{the tracks object from which certain tracks are to be chosen.}

\item{ids}{a vector containing the ids of the tracks that shall be chosen.}
}
\value{
A \emph{tracks} object containing just the elements from the given
\emph{tracks} object which are specified in \code{ids}.
}
\description{
Returns a \emph{tracks} object containing only the specified elements of a given
\emph{tracks} object.
}
\details{
The elements given in \code{ids} are chosen from the list that represents
the \emph{tracks} object and returned, likewise as a \emph{tracks} object.
}

