% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{hotellingsTest}
\alias{hotellingsTest}
\title{Unbiasedness of Motion}
\usage{
hotellingsTest(tracks, dim = c("x", "y"), plot = FALSE)
}
\arguments{
\item{tracks}{the tracks whose biasedness is to be determined.}

\item{dim}{vector with the names of the track's
dimensions that are to be considered. By default c("x", "y").}

\item{plot}{logical indicating whether the scatter of the step's directions,
origin of ordinates (green circle) and the mean of the data points (green
cross) are to be plotted. (In one dimension also the bounds of the
condfidence interval are given.) Plot works only in one or two dimensions.}
}
\description{
Determines the p-value for the motion described by a \emph{tracks} object to
be unbiased.
}
\details{
Computes the displacement vectors of all segments in the tracks
given in \code{tracks}, and determines the p-value for this distribution
using Hotelling's T-square Test
(see \code{\link[DescTools]{HotellingsT2Test}}).
}

