% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{hurstExponent}
\alias{hurstExponent}
\title{Track Hurst Exponent}
\usage{
hurstExponent(x)
}
\arguments{
\item{x}{the input track.}
}
\value{
The corrected empircal Hurst exponent as estimated by
\code{\link[pracma]{hurstexp}}
is returned. If the track has less than two
positions, NA will be returned.
}
\description{
Computes the empirical Hurst exponent of the track.
}
\seealso{
A related measure is the \code{\link{fractalDimension}}.
}

