% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interpolateTrack}
\alias{interpolateTrack}
\title{Interpolate Track Positions}
\usage{
interpolateTrack(x, t, how = "linear")
}
\arguments{
\item{x}{the input track (a matrix or data frame).}

\item{t}{the times at which to approximate track positions. These must lie 
within the interval spanned by the track timepoints.}

\item{how}{specifies how to perform the interpolation. Possible values are 
\code{"linear"} (which uses \code{\link[stats]{approx}} with default values) and
\code{"spline"} (which uses \code{\link[stats]{spline}} with default values).}
}
\description{
Approximates the track positions at given time points using linear interpolation
(via the \code{\link[stats]{approx}} function).
}

