% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.tracks}
\alias{is.tracks}
\title{Check for Class Tracks}
\usage{
is.tracks(x)
}
\arguments{
\item{x}{the input object.}
}
\description{
Verifies whether the input object inherits from S3 class \code{"tracks"}.
}

