% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{maxDisplacement}
\alias{maxDisplacement}
\title{Maximal Displacement}
\usage{
maxDisplacement(x, limits = c(1, nrow(x)))
}
\arguments{
\item{x}{the track whose maximal displacement is to be computed.}

\item{limits}{vector giving the first and last row of the track. Can be used to avoid
computing subtracks explicitly, which is exploited by
\code{\link{aggregate.tracks}}.}
}
\value{
The maximal displacement, a nonnegative number.
}
\description{
Computes the maximal distance of any position on the track from the
starting position.
}

