% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{maxTrackLength}
\alias{maxTrackLength}
\title{The Maximum Number of Points of a Track in a \emph{Tracks} Object}
\usage{
maxTrackLength(tracks)
}
\arguments{
\item{tracks}{the \emph{tracks} object the tracks in which are to be considered.}
}
\value{
Returns the maximum number of rows of a track in the given
\emph{tracks} object.
}
\description{
Determines the maximum number of points that one of the tracks in the given
\emph{tracks} object constists of.
}
\details{
The maximum number of rows of a track in the \emph{tracks} object
is determined.
}

