% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{meanTurningAngle}
\alias{meanTurningAngle}
\title{A Track's Mean Turning Angle}
\usage{
meanTurningAngle(track)
}
\arguments{
\item{track}{the track whose mean turning angle is to be computed.}
}
\value{
The tracks mean turning angle will be returned in degrees.
}
\description{
Computes the mean of all angles between two subsequent segments of the given
track.
}
\details{
Computes the angle between each two subsequent segments in the
track and returns their mean. Angles are metered symmetrically, thus
yielding (degree) values between \eqn{0} and \eqn{180}. (Both a \eqn{90}
degrees left and right turn yield the value 90.)
}

