% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeTrack}
\alias{normalizeTrack}
\title{Normalize a Track}
\usage{
normalizeTrack(track)
}
\arguments{
\item{track}{the track that is to be translated, given in a data frame with
columns \eqn{t}, \eqn{x} and possibly \eqn{y}, and \eqn{z}.}
}
\value{
Retruns the translation of the input track whose starting point is
on the origin of ordinates.
}
\description{
Translates a track such that its starting point is in the origin of ordinates.
}
\details{
Subtracts the coodinates and time of the tracks starting point from
every point on the track.
}

