% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeTracks}
\alias{normalizeTracks}
\title{Normalize Tracks}
\usage{
normalizeTracks(tracks)
}
\arguments{
\item{tracks}{the \code{tracks} object that is to be normalized.}
}
\value{
A \emph{tracks} object with the tracks from the input object translated
such that their starting point is in the origin of ordinates.
}
\description{
Normalizes list of tracks (given in a \code{tracks} object) such that the
starting point of each track is in the origin of ordinates.
}
\details{
The function \code{\link{normalizeTrack}} is applied to each track.
}

