% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{overallDot}
\alias{overallDot}
\title{Dot Product Between Steps at Track Endpoints}
\usage{
overallDot(x, from = 1, to = nrow(x), xdiff = diff(x))
}
\arguments{
\item{x}{the input track. Must have at least 2 rows or an error is thrown.}

\item{from}{index of first row of the track.}

\item{to}{index of last row of the track. 
Both `from` and `to` can be vectors and are useful to avoid extracting subtracks, 
which is exploited by \code{\link{aggregate.tracks}}.}

\item{xdiff}{row differences of x, should be pre-computed if this function is calleed
repeatedly for subtracks of a long track.}
}
\value{
The dot product.
}
\description{
Computes the dot product between the first and last steps of the given track.
}
\examples{
## compute and plot the autocovariance function for the T cell data 
## (assuming isotropy)
with( (aggregate(BCells,overallDot,FUN="mean.se")),{
  plot( mean ~ i, xlab="time", 
  ylab="autocovariance", type="l" )
  segments( i, lower, y1=upper )
} )
}

