% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.tracks}
\alias{plot.tracks}
\title{Plot in 2D Space}
\usage{
\method{plot}{tracks}(x, dims = c("x", "y"), add = F,
  col = order(names(x)), xlab = "X position", ylab = "Y position", ...)
}
\arguments{
\item{x}{the tracks that shall be plotted, in the form of a object of
class "tracks", as returned by e.g. \code{\link{read.tracks.csv}}.}

\item{dims}{a vector giving the dimensions of the track data that shall be
plotted, e.g. \code{c('x','y')} for the \eqn{x} and \eqn{y} dimension.}

\item{add}{boolean value indicating whether the tracks are to be added to the
current plot.}

\item{col}{a specification of the color(s) to be used. This can be a vector
of size \code{length(x)}, where each entry specififes the color for the
corresponding track.}

\item{xlab}{a sting that should be used as a label for the x axis.}

\item{ylab}{a sting that should be used as a label for the y axis.}

\item{...}{additional parameters to be passed to \code{\link[graphics]{plot}}
(in case add=False) or \code{\link[graphics]{points}} (in case add=True),
respectively.}
}
\description{
Plots tracks contained in a "tracks" object into a twodimensional space
pallelel to the data's axes.
}
\details{
One dimension of the data (by default \eqn{y}) is plotted against
another (by default \eqn{x}). The dimesions can be chosen by means of the
parameter \code{dims} and the axes can be labeled accordingly with the aid
of \code{xlab} and \code{ylab}. The color can be set through \code{col}.
If the tracks should be added to an existing plot, \code{add} is to be set
to \code{TRUE}.
}

