% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot3d}
\alias{plot3d}
\title{Plot tracks in 3D}
\usage{
plot3d(tracks, ...)
}
\arguments{
\item{tracks}{the tracks which will be plotted in 3d}

\item{...}{further arguments to be passed on to
\link[scatterplot3d]{scatterplot3d}}
}
\description{
Takes an input tracks object and plots them in 3D using the
\link[scatterplot3d]{scatterplot3d} function.
}

