% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotInParameterSpace}
\alias{plotInParameterSpace}
\title{Plot Tracks in a 1D or 2D Parameter Space}
\usage{
plotInParameterSpace(tracks, measurex, measurey = function(x) {     0 },
  add = FALSE, xlab = deparse(substitute(measurex)),
  ylab = deparse(substitute(measurey)), col = 1, ...)
}
\arguments{
\item{tracks}{a \emph{tracks} object containing the tracks that are to be
plotted}

\item{measurex}{the measure which constitutes the x-dimension of the
parameter space}

\item{measurey}{the measure which constitutes the y-dimension of the
parameter space. By default constant \eqn{0}.}

\item{add}{a logical indicating whether the tracks are to be added to an
existing plot via \code{\link[graphics]{points}}.}

\item{xlab}{label of the x-axis. By default the name of the input function
\code{measurex}.}

\item{ylab}{label of the y-axis. By default the name of the input function
\code{measurey}.}

\item{col}{the color(s) that are to be used, by default color 1 (black),
see \code{\link[graphics]{par}}}

\item{...}{additional parameters to be passed to \code{\link[graphics]{plot}}
(in case add=False) or \code{\link[graphics]{points}} (in case add=True),
respectively.}
}
\description{
Plots the tracks in a \emph{tracks} object into the parameter space of one
or two given measures.
}
\details{
Plots the tracks in the given \emph{tracks} object into the
parameter space of \code{measurex} and \code{measurey}, or of
\code{measurex} and \eqn{0}, if \code{measurey} is not given.
}

