% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotTrackMeasures}
\alias{plotTrackMeasures}
\title{Bivariate Scatterplot of 2 Given Track Measures}
\usage{
plotTrackMeasures(measurex, measurey, x, add = FALSE,
  xlab = deparse(substitute(measurex)),
  ylab = deparse(substitute(measurey)), ...)
}
\arguments{
\item{measurex}{the measure to be shown on the X axis.}

\item{measurey}{the measure to be shown on the Y axis.}

\item{x}{the argument which the measures shall be applied to. Either this
must be of a type which \code{measurex} and \code{measurey} can directly
be applied to, or, if measurex and measurey are applicable to single tracks,
the argument can be a \emph{tracks} object, which the measures will be
applied to using sapply before plotting the result.}

\item{add}{a logical indicating whether the tracks are to be added to an
existing plot via \code{\link[graphics]{points}}.}

\item{xlab}{label of the x-axis. By default the name of the input function
\code{measurex}.}

\item{ylab}{label of the y-axis. By default the name of the input function
\code{measurey}.}

\item{...}{additional parameters to be passed to \code{\link[graphics]{plot}}
(in case add=False) or \code{\link[graphics]{points}} (in case add=True),
respectively.}
}
\description{
Plots the values of two measures applied on given tracks against each
other.
}
\details{
Plots the value of \code{measurey} applied to \code{x} against the
value of  \code{measurey} applied to \code{y}.
}

