% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{read.tracks.csv}
\alias{read.tracks.csv}
\title{Data Input form a CSV File}
\usage{
read.tracks.csv(file, ...)
}
\arguments{
\item{file}{the name of the file which the data are to be read from, a
readable text-mode connection or a complete URL
(see \code{\link[utils]{read.table}}).}

\item{...}{Further arguments to be passed to \code{read.csv}.}
}
\value{
An object of class \emph{tracks} is returned, which is a list of
data frames, each containing the information on one track. The data frames
have a column \eqn{t}, followed by one column for each of the input track's
coordinates.
}
\description{
Reads cell tracks from a CSV file. which contains a number of tracks, each
with a distinct value in the first field 'id', and with at least 3 points
per track, represented by their pos (?), time stamp \eqn{t} and
coordinate(s), \eqn{x}, \eqn{x} and \eqn{y} or \eqn{x}, \eqn{y} and \eqn{z}.
}
\details{
The input file's first four fields are interpreted as \eqn{id},
\eqn{pos}, \eqn{t} and \eqn{x}, respectively, and, if available, the fifth
as \eqn{y} and the sixth as \eqn{z}. The returned object has the class
\emph{tracks}, which is a list of data frames representing the single
tracks and having columns \eqn{t} and \eqn{x}, plus \eqn{y} and \eqn{z}, if
necessary. The tracks' ids are retained in their position in the list, while
the field \eqn{pos} will be unmaintained.
}

