% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeTracks}
\alias{removeTracks}
\title{Remove Tracks from Tracks Object}
\usage{
removeTracks(x, ids)
}
\arguments{
\item{x}{the input tracks object.}

\item{ids}{a vector containing the ids of the tracks that shall be removed.}
}
\description{
Returns a \emph{tracks} object from the tracks with the given IDs are removed.
}

