% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{speed}
\alias{speed}
\title{Mean Track Speed}
\usage{
speed(track)
}
\arguments{
\item{track}{a track.}
}
\value{
an estimate of the track's average speed.
}
\description{
Estimate the mean speed of a track by dividing its
length by its duration.
}
\details{
The track's mean speed is computed by dividing its
\code{\link{trackLength}} by its \code{\link{duration}}.
}
\seealso{
\code{\link{trackLength}}, \code{\link{duration}}
}

