% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{straightness}
\alias{straightness}
\title{Track Straightness}
\usage{
straightness(x)
}
\arguments{
\item{x}{the track whose straightness is to be computed.}
}
\value{
The track's straightness, a value between \eqn{0} and \eqn{1}. 
If the track has \eqn{0}, then 1 is returned.
}
\description{
Computes a track's straightness, i.e., the track's \code{\link{displacement}} 
divided by its \code{\link{trackLength}}.
}
\seealso{
\code{\link{asphericity}} for a different measure of straightness.
}

