% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{subtracks}
\alias{subtracks}
\title{All Subtracks with a Certain Number of Segments for a \emph{Tracks} Object}
\usage{
subtracks(tracks, i, overlap = i - 1)
}
\arguments{
\item{tracks}{the \emph{tracks} object whose subtracks are to be computed.}

\item{i}{the number of segments the subtracks ahll consist of.}

\item{overlap}{the number of segments in which each subtrack shall overlap
its predecessor and successor. Default is \eqn{i - 1}.}
}
\value{
A \emph{tracks} object is returned which contains all the subtracks
of any track in the input \emph{tracks} object that consist of exactly \code{i}
segments and overlap their predecessor in  \code{overlap} segments.
}
\description{
Computes a \emph{tracks} object of all the subtracks of the input
\emph{tracks} object that constist of \code{i} segments.
}
\details{
The function \code{\link{computeSubtracksOfISegments}} is applied
on every track for the given \code{i} and \code{overlap} and the result is
returned as a single \emph{tracks} object.
}

