% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{trackLength}
\alias{trackLength}
\title{Track Length}
\usage{
trackLength(track)
}
\arguments{
\item{track}{The track whose length is to be computed.}
}
\value{
The length.
}
\description{
Estimates the length of the track as the sum of the lengths of its
linear segments. (Note that this is an underestimation).
}
\details{
The track length is computed by summing the Euclidean distances
between consecutive positions.
}

