% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{wrapTrack}
\alias{wrapTrack}
\title{Wrap a Single Track into a \emph{Tracks} Object}
\usage{
wrapTrack(track)
}
\arguments{
\item{track}{the track that is to be wrapped.}
}
\value{
Retruns a \emph{tracks} object containing the input \code{track}
named "1".
}
\description{
Makes a \emph{tracks} object out of a single track.
}
\details{
A list containing only the input \code{track} is returned. It is
assigned the name "1". This list is returned as a \emph{tracks}
object.
}

