% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{as.data.frame.tracks}
\alias{as.data.frame.tracks}
\title{Convert from Tracks to Data Frame}
\usage{
\method{as.data.frame}{tracks}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{the  \code{tracks} object to be coerced to a data frame.}

\item{row.names}{NULL or a character vector giving the row names for the
data frame.  Missing values are not allowed.}

\item{optional}{logical. Column names are always assigned to the resulting
data frame regardless of the setting of this option.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
Returns one data frame containing all data of the indiviual
tracks from the input with a prepended column named "id" which contains
each track's name in the list of the \emph{tracks} object.
}
\description{
Converts a \code{tracks} object into a data frame.
}
\details{
Returns one data frame containing all indiviual
tracks from the input with a prepended column "id" containing
each track's name in the list of the \emph{tracks} object.
}

