% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{as.tracks.list}
\alias{as.tracks.list}
\title{Convert from List to Tracks}
\usage{
\method{as.tracks}{list}(x, ...)
}
\arguments{
\item{x}{the list that is to be converted into a \emph{tracks} object.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
the coerced object of S3 class \code{tracks} (i.e., a list
 containing a matrix for each track)
}
\description{
"Converts" a list into a \code{tracks} object by settings its
S3 class to \code{tracks}.
}

