% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{asphericity}
\alias{asphericity}
\title{A Track's Aspericity}
\usage{
asphericity(track)
}
\arguments{
\item{track}{the track whose asphericity is to be computed. Note that the
track must be in a two- or threedimensional space, thus the data frames
containing the track's points must contain an \eqn{x} and \eqn{y} value.}
}
\value{
The track's asphericity, a vaule between \eqn{0} and \eqn{1}, will
be returned for two- or threedimensional tracks, otherwise \code{NA}.
}
\description{
Computes the asphericity of the track's points scatter plot.
}
\details{
Computes the asphericity of a track, i.e. the scatter plot of
its points, by considering the length of its principal components,
see References. The value will be from the interval \eqn{[0,1]},
with high values indicating track straightness. The asphericity is a
measure for the track's straightness. Note that the asphericity is only
defined for tracks in two or three dimensions.
}
\references{
\cite{Mokhtari et al.: Automated Characterization and
Parameter-Free Classification of Cell Tracks Based on Local Migration
Behavior, PLoS ONE, 2013}
}

