% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{asphericity}
\alias{asphericity}
\title{Track Aspericity}
\usage{
asphericity(x)
}
\arguments{
\item{x}{input track, which must have two or more spatial dimensions.}
}
\value{
The track's asphericity, a value between \eqn{0} and \eqn{1}, is
returned for two- or higher dimensional tracks. Otherwise the return value is \code{NA}.
}
\description{
Computes the asphericity of the set of positions on the track
via the length of its principal components
(Mokhtari et al, 2013).  Asphericity is a
measure for the track's straightness: it is a number between 0 and 1, with
higher values indicating straighter tracks.
In that sense it is similar to \code{\link{straightness}}, however, the asphericity is
not affected by back-and-forth motion of the object.
}
\details{
We define the asphericity of every track with two or fewer positions to be 1.
This is true even in the case where two positions are identical, which is motivated
by considering
that situation as the limit of a process in which two points move towards each other.
}
\references{
Zeinab Mokhtari, Franziska Mech, Carolin Zitzmann, Mike Hasenberg, Matthias Gunzer
and Marc Thilo Figge (2013), Automated Characterization and
Parameter--Free Classification of Cell Tracks Based on Local Migration
Behavior. \emph{PLoS ONE} \bold{8}(12), e80808. doi:10.1371/journal.pone.0080808
}
\seealso{
\code{\link{straightness}}
}

