% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{boundingBox}
\alias{boundingBox}
\title{The Bounding Box of a \emph{Tracks} Object}
\usage{
boundingBox(tracks)
}
\arguments{
\item{tracks}{the \emph{tracks} object whose bounding box is to be computed}
}
\value{
Returns a matrix with two rows and \eqn{d} columns, where \eqn{d} is
the number of dimensions of the tracks. The first row contains the minimum
and the second row the maximum value of any track in the dimension given by
the column.
}
\description{
Computes minimum and maximum value that a track has in each dimension
}
\details{
In each of the tracks' dimensions the minimum and maximum value
that any track from the input has is computed and returned. It is assumed
that all the tracks have the same number of dimensions.
}

