% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacementRatio}
\alias{displacementRatio}
\title{A Track's Displacement Ratio}
\usage{
displacementRatio(track)
}
\arguments{
\item{track}{the track whose displacement ratio is to be computed.}
}
\value{
The track's displacement ratio, a vaule between \eqn{0} and \eqn{1},
will be returned. If the track has a maximal displacement of \eqn{0}, then
NaN is returned.
}
\description{
Computes a track's displacement ratio, i.d. the track's (current)
displacement divided by its maximal displacement.
}
\details{
Computes the displacement ratio of a track, i.e. the ratio of the
track's (current) displacement and its maximum displacement. The value will
be from the interval \eqn{[0,1]}, or 1 if the track has a maximal
displacement of 0.
}
\references{
\cite{Mokhtari et al.: Automated Characterization and
Parameter--Free Classification of Cell Tracks Based on Local Migration
Behavior, PLoS ONE, 2013}
}
\seealso{
\code{\link{displacement}}, \code{\link{maxDisplacement}}
}

