% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacementRatio}
\alias{displacementRatio}
\title{Track Displacement Ratio}
\usage{
displacementRatio(track)
}
\arguments{
\item{track}{the track whose displacement ratio is to be computed.}
}
\value{
The track's displacement ratio, a vaule between \eqn{0} and \eqn{1},
will be returned. If the track has a maximal displacement of \eqn{0}, then
NaN is returned.
}
\description{
Computes a track's displacement ratio, i.d. the track's (current)
displacement divided by its maximal displacement.
}
\details{
Computes the displacement ratio of a track, i.e. the ratio of the
track's (current) displacement and its maximum displacement. The value will
be from the interval \eqn{[0,1]}, or 1 if the track has a maximal
displacement of 0.
}
\references{
Zeinab Mokhtari, Franziska Mech, Carolin Zitzmann, Mike Hasenberg, Matthias Gunzer
and Marc Thilo Figge (2013), Automated Characterization and
Parameter--Free Classification of Cell Tracks Based on Local Migration
Behavior. \emph{PLoS ONE} \bold{8}(12), e80808. doi:10.1371/journal.pone.0080808
}
\seealso{
\code{\link{displacement}}, \code{\link{maxDisplacement}}
}

