% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacementVector}
\alias{displacementVector}
\title{Vector Between Track Endpoints}
\usage{
displacementVector(x)
}
\arguments{
\item{x}{the track whose displacement vector is to be computed.}
}
\value{
The displacement vector, which has as many dimensions as the input tracks.
}
\description{
Computes the vector between the track endpoints.
}

