% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{displacementVector}
\alias{displacementVector}
\title{Vector Between Track Endpoints}
\usage{
displacementVector(track)
}
\arguments{
\item{track}{the track whose displacement vector is to be computed.}
}
\value{
The difference between the coordinates of the track's end point and
those of its starting point is returned.
}
\description{
Computes the coordinates of the vector between that tracks' start and end points.
}
\details{
The coordinates of the tracks's starting point are subtracted
from its end point's coordinates.
}

