% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{duration}
\alias{duration}
\title{Track Duration}
\usage{
duration(track)
}
\arguments{
\item{track}{a track.}
}
\value{
Returns a tracks duration.
}
\description{
Returns the duration of a track.
}
\details{
Computes the track's duration by subtracting the
time of the its first point from the time of its last.
}

