% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{forEveryPrefix}
\alias{forEveryPrefix}
\title{Measure for Every Prefix of a Track}
\usage{
forEveryPrefix(measure, min.length = 1)
}
\arguments{
\item{measure}{the measure that is to be computed.}

\item{min.length}{minimal number of points in a prefix the measure is to be
applied to. Default is \eqn{1}.}
}
\value{
A function that applies the measure to every prefix of length at
least \code{min.length} is returned.
}
\description{
Returns a function that applies the input measure to each of a tracks prefixes
of a given minimal length.
}
\details{
A function that applies the measure to every prefix of length at
least \code{min.length} is returned. This
function applies the measure to every prefix of the input track, starting
with the prefix that consists of \code{min.length} points. The resulting
values are returned in a vector, sorted in ascending order by the
prefix length.
}

