% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getTracks}
\alias{getTracks}
\title{Get a Subset of Tracks}
\usage{
getTracks(x, ids)
}
\arguments{
\item{x}{the input tracks object.}

\item{ids}{a vector containing the ids of the tracks that shall be chosen.}
}
\value{
the modified \code{tracks} object.
}
\description{
Returns a \code{tracks} object containing only the specified elements of a given
\code{tracks} object.
}
\details{
the elements given in \code{ids} are chosen from the list that represents
the \code{tracks} object and returned, likewise as a \code{tracks} object.
}

