% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{hurstExponent}
\alias{hurstExponent}
\title{Track Hurst Exponents}
\usage{
hurstExponent(x)
}
\arguments{
\item{x}{the track whose Hurst exponent is to be computed.}
}
\value{
A vector with the Hurst exponent for each of the tracks dimension
is returned. If the track has an uneven number of points, the last one is
discarded, if it has less than six points, NA will be returned.
}
\description{
Computes the Hurst exponent of each of the track's dimensions.
}
\details{
If the track includes at least six and an even number of points,
for each of the track's dimensions, the empirical Hurst exponent is
determined using the function \code{\link[pracma]{hurstexp}} from the
\code{pracma} package and returned as part of a vector.
If the track consists of less than six points, NA is returned, if it
comprises an uneven number of points, the last one is discarded.
}
\seealso{
\code{\link{fractalDimension}}
}

