% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{maxDisplacement}
\alias{maxDisplacement}
\title{Maximal Displacement}
\usage{
maxDisplacement(track, limits = c(1, nrow(track)))
}
\arguments{
\item{track}{the track whose maximal displacement is to be computed.}

\item{limits}{Vector giving the first and last row of the track. Can be used to avoid
extracting subtracks, which is exploited e.g. by
\code{\link{aggregate.tracks}}.}
}
\value{
The maximal displacement.
}
\description{
Computes the maximal distance of any position on the track from the
starting position.
}
\details{
Computes the maximum over all pairwise Euclidean distances between
the track's starting point and another point on the track.
}

