% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{maxDisplacement}
\alias{maxDisplacement}
\title{Maximal Displacement}
\usage{
maxDisplacement(x)
}
\arguments{
\item{x}{the track whose maximal displacement is to be computed.}
}
\value{
The maximal displacement, a nonnegative number.
}
\description{
Computes the maximal distance of any position on the track from the 
starting position.
}

