% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{outreachRatio}
\alias{outreachRatio}
\title{A Track's Outreach Ratio}
\usage{
outreachRatio(track)
}
\arguments{
\item{track}{the track whose outreach ratio is to be computed.}
}
\value{
The track's outreach ratio, a vaule between \eqn{0} and \eqn{1},
will be returned. If the track has length \eqn{0}, then NaN is returned.
}
\description{
Computes a track's outreach ratio, i.d. the track's maximal
displacement divided by its length.
}
\details{
Computes the outreach ratio of a track, i.e. the ratio of the
track's maximal displacement and length. The value will be from the interval
\eqn{[0,1]}, or 1 if the track has length \eqn{0}.
}
\references{
\cite{Mokhtari et al.: Automated Characterization and
Parameter--Free Classification of Cell Tracks Based on Local Migration
Behavior, PLoS ONE 2013 Dec 6;8(12):e80808, 2013}
}
\seealso{
\code{\link{maxDisplacement}}, \code{\link{trackLength}}
}

