% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{overallAngle}
\alias{overallAngle}
\title{A Track's Overall Turning Angle}
\usage{
overallAngle(track, limits = c(1, nrow(track)))
}
\arguments{
\item{track}{the track whose overall turning angle is to be computed.}

\item{limits}{Vector giving the first and last row of the track. Can be used to avoid
extracting subtracks, which is exploited e.g. by
\code{\link{aggregateSubtrackMeasures}}.}
}
\value{
The track's overall turning angle in radians.
}
\description{
Computes the angle between the first and the last segment of the given track.
}
\details{
Computes the angle between the vectors representing the track's
first and last segment, respectively, i.e. the overall turning angle.
Angles are measured symmetrically, thus yielding (degree) values between
0 and 180. (Both a 90 degrees left and right turn yield the value 90.)
}

