% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{overallAngle}
\alias{overallAngle}
\title{Angle Between Steps at Track Endpoints}
\usage{
overallAngle(x, from = 1, to = nrow(x), xdiff = diff(x))
}
\arguments{
\item{x}{the track whose overall turning angle is to be computed.}

\item{from}{index of first row of the track.}

\item{to}{index of last row of the track.
Both `from` and `to` can be vectors and are useful to avoid extracting subtracks,
which is exploited by \code{\link{aggregate.tracks}}.}

\item{xdiff}{row differences of x, should be pre-computed if this function is calleed
repeatedly for subtracks of a long track.}
}
\value{
The track's overall turning angle in radians.
}
\description{
Computes the angle between the first and the last segment of the given track.
}
\details{
Computes the angle between the vectors representing the track's
first and last segment, respectively, i.e. the overall turning angle.
Angles are measured symmetrically, thus the return values range from 0 to 180 degrees.
For instance, both a 90 degrees left and right turn yield the value 90.
}
\examples{
## show a turning angle plot with error bars for the T cell data.
with( (aggregate(BCells,overallDot,FUN="mean.se",na.rm=TRUE)),{
  plot( mean ~ i, xlab="time step",
  	ylab="turning angle (rad)", type="l" )
  segments( i, lower, y1=upper )
} )
}

