% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{overallDot}
\alias{overallDot}
\title{Dot Product Between Steps at Track Endpoints}
\usage{
overallDot(x, limits = c(1, nrow(x)))
}
\arguments{
\item{x}{the input track.}

\item{limits}{vector giving the first and last row of the track. Can be used to avoid
extracting subtracks, which is exploited e.g. by
\code{\link{aggregate.tracks}}.}
}
\value{
the dot product.
}
\description{
Computes the dot product between the first and last steps of the given track.
}
\examples{
## compute and plot the autocovariance function for the T cell data (assuming isotropy)
with( (aggregate(BCells,overallDot,FUN="mean.se",na.rm=TRUE)),{
  plot( cos(mean) ~ i, xlab="time",
  ylab="correlation of orientation", type="l" )
  segments( i, cos(lower), y1=cos(upper) )
} )
}

