% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-set-transformations.R
\name{projectDimensions}
\alias{projectDimensions}
\title{Extract Subset of Dimensions}
\usage{
projectDimensions(tracks, dims = c("x", "y"))
}
\arguments{
\item{tracks}{the \emph{tracks} onject whose tracks are to be projected.}

\item{dims}{a character vector giving the dimensions to extract from each track.}
}
\value{
A tracks object is returned that contains only those dimensions
of the input \code{tracks} that are given in \code{dims}.
}
\description{
Projects tracks in a \emph{tracks} object onto the given dimensions.
}

