% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-set-transformations.R
\name{projectDimensions}
\alias{projectDimensions}
\title{Extract Subset of Dimensions}
\usage{
projectDimensions(x, dims = c("x", "y"))
}
\arguments{
\item{x}{the input tracks object.}

\item{dims}{a character vector giving the dimensions to extract from each track.}
}
\value{
A tracks object is returned that contains only those dimensions
of the input \code{tracks} that are given in \code{dims}.
}
\description{
Projects tracks onto the given dimensions.
}
\examples{
## Compare 2D and 3D speeds
speed.2D <- mean( sapply( subtracks( projectDimensions( TCells, c("x","z") ), 2 ), speed ) )
speed.3D <- mean( sapply( TCells, speed ) )
}

