% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{speed}
\alias{speed}
\title{Mean Track Speed}
\usage{
speed(x)
}
\arguments{
\item{x}{a track.}
}
\value{
An estimate of the track's average speed, a nonnegative number.
}
\description{
Estimate the mean instantaneous speed of a track by dividing its
\code{\link{trackLength}} by its \code{\link{duration}}.
}

