% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{squareDisplacement}
\alias{squareDisplacement}
\title{Squared Displacement}
\usage{
squareDisplacement(x, from = 1, to = nrow(x))
}
\arguments{
\item{x}{the input track.}

\item{from}{index of first row of the track.}

\item{to}{index of last row of the track. 
Both `from` and `to` can be vectors and are useful to avoid extracting subtracks, 
which is exploited by \code{\link{aggregate.tracks}}.}
}
\value{
The track's square displacement will be returned.
}
\description{
Computes the squared distance between the track's first and final positions.
}

