% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{squareDisplacement}
\alias{squareDisplacement}
\title{Square Displacement}
\usage{
squareDisplacement(x, limits = c(1, nrow(x)))
}
\arguments{
\item{x}{the track whose displacement is to be computed.}

\item{limits}{Vector giving the first and last row of the track. Can be used to avoid
extracting subtracks, which is exploited e.g. by
\code{\link{aggregate.tracks}}.}
}
\value{
The track's square displacement will be returned.
}
\description{
Computes the squared distance between the track's first and final positions.
}

