% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{squareDisplacement}
\alias{squareDisplacement}
\title{Square Displacement}
\usage{
squareDisplacement(track, limits = c(1, nrow(track)))
}
\arguments{
\item{track}{the track whose displacement is to be computed.}

\item{limits}{Vector giving the first and last row of the track. Can be used to avoid
extracting subtracks, which is exploited e.g. by
\code{\link{aggregate.tracks}}.}
}
\value{
The track's square displacement will be returned.
}
\description{
Computes the squared distance between the track's first and final positions.
}
\details{
Computes the squared Euclidean distance between the track's start
and end point, by not extrakting the square root of the squared distances on
the single dimensions.
}

