% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{staggered}
\alias{staggered}
\title{Staggered Version of a Function}
\usage{
staggered(measure, ...)
}
\arguments{
\item{measure}{the measure that shall be computed in a staggered fashion.}

\item{...}{further parameters for staggered computation of the measure
(see \code{\link{computeStaggered}}).}
}
\value{
Returns a funtion of a cell track (in a data frame) that computes
the given measure in a staggered fashion on that track.
}
\description{
Returns the staggered version of a track measure.
}
\details{
Returns a funtion of a cell track (in a data frame) that computes
the given measure in a staggered fashion on that track.
}

