% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track-measures.R
\name{straightness}
\alias{straightness}
\title{A Track's Straightness}
\usage{
straightness(track)
}
\arguments{
\item{track}{the track whose straightness is to be computed.}
}
\value{
The track's straightness, a value between \eqn{0} and \eqn{1}, will
be returned. If the track has length \eqn{0}, then NaN is returned.
}
\description{
Computes a track's straightness, i.d. the track's displacement divided by
its length.
}
\details{
Computes the straightness of a track, i.e. the ratio of the
track's displacement and length. The value will be from the interval
\eqn{[0,1]}, or 1 if the track has length \eqn{0}.
}
\seealso{
\code{\link{displacement}}, \code{\link{trackLength}}
}

