% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-measures.R
\name{trackLength}
\alias{trackLength}
\title{Track Length}
\usage{
trackLength(x)
}
\arguments{
\item{x}{The track whose length is to be computed.}
}
\value{
The length.
}
\description{
Estimates the length of the track as the sum of the lengths of its
linear segments.
}
\details{
This amounts to linar interpoplation, which really gives a lower
bound on the track length and is therefore usually an underestimation.
}

